refactor: lottoQR 카메라 인식 코드 refactoring
