+++ LottoDiary/LottoDiary/Sources/App/SceneDelegate.swift
@@ -11,6 +11,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
 
+    // https://blog.naver.com/PostView.nhn?blogId=psj9102&logNo=222094915075&categoryNo=73&parentCategoryNo=0&viewDate=&currentPage=1&postListTopCurrentPage=1&from=postView
+    var goalAmountSetting = false
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
@@ -18,7 +20,12 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let scene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: scene)
-        window?.rootViewController = IntroViewController()
+        
+        if goalAmountSetting {
+            window?.rootViewController = UINavigationController(rootViewController: TabBarController())
+        } else {
+            window?.rootViewController = UINavigationController(rootViewController: MyInfomationSettingViewController())
+        }
         window?.makeKeyAndVisible()
     }
