+++ LottoDiary/LottoDiary/Sources/Present/LottoQRManager/LottoResultModels.swift
