+++ LottoDiary/LottoDiary/Sources/Present/NumberDraw/Models/LottoData.swift
@@ -19,3 +19,33 @@ extension LottoData {
         LottoData(turnNumber: 1035, numbers: [15,16,17,18,19,20,21])
     ]
 }
+
+
+
+/*
+ import UIKit
+
+ let standardDate = "2022-12-24"
+ var standardRound = 1047
+ let today = "2023-01-06"
+
+ let dataFormatter = DateFormatter()
+ dataFormatter.dateFormat = "yyyy-MM-dd"
+
+ let startDate = dataFormatter.date(from: standardDate)
+ let endDate = dataFormatter.date(from: today)
+
+ let term = Calendar.current.dateComponents([.day], from: startDate!, to: endDate!)
+
+ if term.day! > 7 {
+     standardRound += term.day! / 7
+ } else {
+     if term.day! == 7 {
+          standardRound += 1
+      }
+ }
+ //print(term.day! / 7)
+
+ print(standardRound)
+
+ */
