--- LottoDiary/LottoDiary/Sources/Present/RandomLottoNumberViewController.swift
@@ -1,16 +1,0 @@
-//
-//  RandomLottoNumberViewController.swift
-//  LottoDiary
-//
-//  Created by 송선화 on 2022/12/02.
-//
-
-import UIKit
-
-final class RandomLottoNumberViewController: UIViewController {
-
-    override func viewDidLoad() {
-      super.viewDidLoad()
-      view.backgroundColor = .blue
-  }
-}
